% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/625rpackage-package.R
\name{biostat625rpackage}
\alias{biostat625rpackage}
\title{biostat625rpackage: A package containing functions and Rcpp code for linear regression modeling}
\description{
biostat625rpackage: A package containing functions and Rcpp code for linear regression modeling
}
