% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/625_lm.R
\name{lm_625}
\alias{lm_625}
\title{Linear Regression Model}
\usage{
lm_625(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictor variables.}

\item{y}{A numeric vector of response variables.}
}
\value{
is a list summarizing the fitted model results.
}
\description{
This function fits a simple linear regression model to the provided data,
using basic matrix algebra. It returns coefficients, fitted values, residuals,
R-squared, adjusted R-squared, standard error, F-statistic, and degrees of freedom.
}
\examples{
data(mtcars)
x <- as.matrix(mtcars[, c("wt", "hp", "cyl")])
y <- mtcars$mpg
results <- lm_625(x, y)
print(results)

}
